TECH-334: Migrate Breadcrumb to typescript
